+++ resources/views/ContractsRequests/infoSign.blade.php
@@ -2,21 +2,23 @@
     <div class="modal-dialog modal-lg" role="document">
         <div class="modal-content">
             <div class="modal-header">
-                <h5 class="modal-title" id="applySignatureModalTitle">Contract 2<div class="badge badge-warning ml-2">NOT FINALIZED</div></h5>
-                <button type="button" class="close waves-effect" data-dismiss="modal" aria-label="Close">
+                <h5 class="modal-title" id="applySignatureModalTitle"><div class="badge badge-warning">Finalizing...</div></h5>
+                <button type="button" class="close waves-effect modalClose" data-dismiss="modal" aria-label="Close">
                     <span aria-hidden="true">&times;</span>
                 </button>
             </div>
             <div class="modal-body">
                 <div class="alert alert-info mt-2">
                     <i class="bigIcon fas fa-info-circle mr-2 mt-1"></i>
-                    You must sign in the canvas first before accepting the contract.
+                    You must sign in the canvas first before finalizing the contract.
                 </div>
                 <div class="signCanvas"></div>
                 <button class="clearCanvas btn btn-primary btn-sm waves-effect float-left">Clear Canvas</button> 
             </div>
             <div class="modal-footer">
-                <button onclick="acceptContractQuotation()" class="btn btn-primary waves-effect btnButtons" data-toggle="modal" data-target="#editContractInfo">Accept Contract</button>
+                <button onclick="createActiveContract({{$companyAccepted->intContractListID}})" class="btn waves-effect btn-success">
+                    Finalize Contract
+                </button>
             </div>
         </div>
     </div>