+++ resources/views/ContractsRequests/styles.blade.php
@@ -1,2 +1,3 @@
 
     <link rel="stylesheet" href="/others/stisla_admin_v1.0.0/dist/css/contractRequests/contractRequests.css">
+    <link rel="stylesheet" href="/others/stisla_admin_v1.0.0/dist/modules/jquery-signature/css/jquery.signature.css">