+++ resources/views/Hauling/list.blade.php
@@ -19,6 +19,27 @@
                     <div class="tab-content" id="pills-tabContent">
                         <div class="tab-pane fade show active" id="pillsPending" role="tabpanel" aria-labelledby="pillsPending-tab">
                             <div class="row">
+<<<<<<< HEAD
+                                <div class="col-12 col-sm-12 col-lg-6">
+                                    <div class="card card-sm-2 card-primary border-primary pendingCards">
+                                        <div class="card-icon">
+                                            <i class="ion ion-android-boat text-primary"></i>
+                                        </div>
+                                        <div class="card-header">
+                                            <h4 class="text-primary mb-2">Job Order # 17</h4>
+                                        </div>
+                                        <div class="card-body">
+                                            <h5>Consignee Name</h5>
+                                        </div>
+                                        <div class="card-footer mt-2">
+                                            <a href="#" data-toggle="modal" data-target="#moreInfoModal">More Info <i class="ion ion-ios-arrow-right"></i></a>
+                                            <button class="btn btn-danger btn-sm text-center float-right ml-2 waves-effect btnEnd">End</button>
+                                            <button class="btn btn-primary btn-sm text-center float-right ml-2 waves-effect btnUL" data-toggle="modal" data-target="#updateLoc">Update Location</button>
+                                            <button class="btn btn-primary btn-sm text-center float-right waves-effect btnStart">Start</button>
+                                        </div>
+                                    </div>
+                                </div>
+=======
                                 @if(count($joborder) > 0)
                                     @foreach($joborder as $joborder)
                                         <div class="col-12 col-sm-12 col-lg-6">
@@ -46,10 +67,32 @@
                                 @else
                                     NO Results
                                 @endif
+>>>>>>> 20c2916bc31bceead8d8bc983914614343f85f0a
                             </div>
                         </div>
                         <div class="tab-pane fade" id="pillsActive" role="tabpanel" aria-labelledby="pillsActive-tab">
                             <div class="row">
+<<<<<<< HEAD
+                                <div class="col-12 col-sm-12 col-lg-6">
+                                    <div class="card card-sm-2 card-primary border-primary activeCards">
+                                        <div class="card-icon">
+                                            <i class="ion ion-android-boat text-primary"></i>
+                                        </div>
+                                        <div class="card-header">
+                                            <h4 class="text-primary mb-2">Job Order # 17</h4>
+                                        </div>
+                                        <div class="card-body">
+                                            <h5>Consignee Name</h5>
+                                        </div>
+                                        <div class="card-footer mt-2">
+                                            <a href="#" data-toggle="modal" data-target="#moreInfoModal">More Info <i class="ion ion-ios-arrow-right"></i></a>
+                                            <button class="btn btn-danger btn-sm text-center float-right ml-2 waves-effect btnEnd">End</button>
+                                            <button class="btn btn-primary btn-sm text-center float-right ml-2 waves-effect" data-toggle="modal" data-target="#updateLoc">Update Location</button>
+                                            <button class="btn btn-primary btn-sm text-center float-right waves-effect btnStart">Start</button>
+                                        </div>
+                                    </div>
+                                </div>
+=======
                                 @if(count($ongoingjob)>0)
                                     @foreach($ongoingjob as $ongoingjob)
                                         <div class="col-12 col-sm-12 col-lg-6">
@@ -75,6 +118,7 @@
                                 @else
                                     No Results Found
                                 @endif
+>>>>>>> 20c2916bc31bceead8d8bc983914614343f85f0a
                             </div>
                         </div>
                     </div>